+++ R/aaa.R
@@ -3,3 +3,4 @@ the$server_processes <- list()
 the$http_allowed_origins <- character()
 the$http_trusted_hosts <- character()
 the$http_shared_secret <- NULL
+the$socket_secret <- NULL