+++ tests/testthat/test-tools.R
@@ -215,8 +215,11 @@ test_that("tools/list gates top-level title on negotiated protocol version", {
 })
 
 test_that("list_r_sessions() filters out integer error codes", {
+  local_socket_secret()
   local_mocked_bindings(
-    collect_aio_ = function(x) list("1: /home/user/myproject (RStudio)", 5L),
+    collect_aio_ = function(x) {
+      list(mac_seal(charToRaw("1: /home/user/myproject (RStudio)")), 5L)
+    },
     socket = function(...) structure(list(), class = "nanoSocket"),
     monitor = function(...) structure(list(), class = "nanoMonitor"),
     dial = function(...) 0L,